requirements.txt
pyautogui
Pillow
opencv-python
pytesseract
pynput
python-dotenv
